fix env handling
