+++ main.py
@@ -1,19 +1,12 @@
 from flask import Flask
 from flask_session import Session
 from flask_cors import CORS
-import dotenv
 from apis import api
-import os
-
-def generate_new_secret_key():
-    key = os.urandom(24).hex()
-    return key 
-
-dotenv.load_dotenv()
+import config as app_config
 
 app = Flask(__name__)
 
-app.config['SECRET_KEY']=generate_new_secret_key()
+app.config['SECRET_KEY']=app_config.config.SECRET_KEY
 app.config['SESSION_PERMANENT'] = True
 app.config.update(SESSION_COOKIE_SAMESITE="None", SESSION_COOKIE_SECURE=True)
 
@@ -31,8 +24,7 @@ def health():
     return "I'm healthy", 200
 
 if __name__ == '__main__':
-    print('FCO FCO FCO ', os.getenv('FLASK_ENV'))
-    if os.getenv('FLASK_ENV') != 'development':
+    if app_config.config.ENV != 'development':
         from waitress import serve
         serve(app, host="0.0.0.0", port=8080)
     else:
